organise and document hooks
