+++ wp-content/plugins/vf-gutenberg/blocks/hooks/use-vf-defaults.js
@@ -1,0 +1,23 @@
+/**
+ * Return default Gutenberg block settings for a VF block
+ */
+const {__} = wp.i18n;
+
+const useVFDefaults = () => ({
+  keywords: [__('VF'), __('Visual Framework')],
+  attributes: {
+    ver: {
+      type: 'integer'
+    }
+  },
+  supports: {
+    align: false,
+    className: false,
+    customClassName: false,
+    html: false
+  },
+  edit: () => null,
+  save: () => null
+});
+
+export default useVFDefaults;
