+++ wp-content/plugins/vf-gutenberg/blocks/hooks/use-vf-gutenberg.js
@@ -1,0 +1,23 @@
+/**
+ * Return VF Gutenberg settings provided by the `vfGutenberg` global object
+ * `wp_localize_script` defines this when enqueueing "vf-blocks.js"
+ */
+
+// Default properties
+const vfGutenberg = {
+  postId: 0,
+  nonce: '',
+  plugins: {}
+};
+
+const useVFGutenberg = () => {
+  const vf = window.vfGutenberg || {};
+  for (let [key, value] of Object.entries(vfGutenberg)) {
+    if (!vf.hasOwnProperty(key)) {
+      vf[key] = value;
+    }
+  }
+  return vf;
+};
+
+export default useVFGutenberg;
