+++ wp-content/plugins/vf-gutenberg/blocks/vf-blocks.jsx
@@ -1,7 +1,8 @@
 /**!
  * VF Gutenberg blocks
  */
-import {useVFGutenberg, useVFPluginSettings} from './hooks';
+import useVFGutenberg from './hooks/use-vf-gutenberg';
+import useVFSettings from './hooks/use-vf-settings';
 
 // Import Visual Framework core component settings
 import vfButton from './vf-core/vf-button';
@@ -14,6 +15,6 @@ registerBlockType(vfButton.name, vfButton);
 // Register VF Plugin blocks from "localized" global settings
 const {plugins} = useVFGutenberg();
 for (const [key, plugin] of Object.entries(plugins)) {
-  const settings = useVFPluginSettings(key, plugin.title);
+  const settings = useVFSettings(key, plugin.title);
   registerBlockType(key, settings);
 }
